build(workspace)!: migrate to Rust edition 2024
Workspace configuration
- Set edition = "2024", resolver = "3", and rust-version = "1.96" across the workspace
- Introduce [workspace.package] inheritance for license, repository, categories, edition, and MSRV
- Hoist internal proptest/proptest-macro path deps into [workspace.dependencies] and inherit them in member crates

Dependency upgrades
- Bump bitflags, num-traits, regex, rusty-fork, syn, and tempfile to current releases
- Add serde_json as a workspace dependency and proptest-derive dev-dependency
- Move message-io transport features (tcp/udp/websocket) to the state-machine dev-dependency that uses them

Code migration
- Apply nightly rustfmt and clippy auto-fixes for edition 2024 style and diagnostics
- Rework the num.rs sampling macros to thread a sample mode instead of relying on turbofished generics
- Adapt string.rs, test_runner/config.rs, and test_runner/rng.rs to the upgraded APIs

Docs and tooling
- Bump the CHANGELOG Breaking Changes MSRV note to 1.96 in all four crates
- Document the MSRV policy in book/src/intro.md and raise the pinned CI toolchain to 1.96.0
- Add AGENTS.md guidance with a CLAUDE.md stub importing it

+++ proptest/src/arbitrary/primitives.rs
@@ -12,10 +12,12 @@
 use crate::bool;
 use crate::char;
 use crate::num::{
-    f32, f64, i128, i16, i32, i64, i8, isize, u128, u16, u32, u64, u8, usize,
+    f32, f64, i8, i16, i32, i64, i128, isize, u8, u16, u32, u64, u128, usize,
 };
 
-arbitrary!(bool, i8, i16, i32, i64, isize, u8, u16, u32, u64, usize, i128, u128);
+arbitrary!(
+    bool, i8, i16, i32, i64, isize, u8, u16, u32, u64, usize, i128, u128
+);
 
 // Note that for floating point types we limit the space since a lot of code
 // isn't prepared for (and is not intended to be) things like NaN and infinity.
